+++ molly/compat/thread_coroutine.lua
@@ -38,6 +38,14 @@ local function scheduler()
                 res = {coroutine.resume(co, thread.thread_id,
                                        unpack(thread['func_args']))}
             end
+            if res[1] == false then
+                -- A thread terminated with an error. Drop the remaining
+                -- threads (fail-stop) and report the error to a caller.
+                for i = table.getn(threads), 1, -1 do
+                    table.remove(threads, i)
+                end
+                return false, tostring(res[2])
+            end
             thread['last_result'] = res
         end
         if coroutine.status(co) == 'dead' then